github/requirements.txt
python-dotenv==0.21.0
requests==2.28.1
pydantic==1.10.2
psycopg2-binary==2.9.3
SQLAlchemy==2.0.30
logstash-formatter==0.5.17
python-json-logger==2.0.7

